re-ordered color-utils and extracted theming helper functions to their own file
