+++ src/shared/index.ts
@@ -1,6 +1,7 @@
 import Utils from './utils';
 
 export * from './color-utils';
+export * from './theme-utils';
 export * from './constants';
 export * from './types';
